# Remove string dequoting from `KToken`