--- a/k-files/string-rewrite.k
+++ b/k-files/string-rewrite.k
@@ -1,0 +1,14 @@
+module STRING-REWRITE-SYNTAX
+    imports STRING-SYNTAX
+endmodule
+
+module STRING-REWRITE
+    imports STRING-REWRITE-SYNTAX
+    imports STRING
+
+    configuration <k> $PGM:String </k>
+                  <s> "" </s>
+
+    rule <k> S => . ... </k>
+         <s> _ => S     </s>
+endmodule