--- a/pyproject.toml
+++ b/pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "pyk"
-version = "0.1.232"
+version = "0.1.233"
 description = ""
 authors = [
     "Runtime Verification, Inc. <contact@runtimeverification.com>",