~Blocked on: https://github.com/runtimeverification/haskell-backend/issues/3563~ **Edit**: skipping tests

Fixes #140
Fixes #314

Removes string dequoting from `KToken`, which now occurs on KAST-to-KORE conversion.