@tothtamas28 2023-04-07
@ehildenb, some newly added integration tests fail due to https://github.com/runtimeverification/haskell-backend/issues/3563. Should I just disable them and proceed with this PR?