@ehildenb 2023-04-12
@tothtamas28 please also test this with KEVM Foundry integration locally, and make sure it works prior to merge. If the tests there aren't passing, then something is going wrong with the bytes encoding choices.