@ehildenb 2023-04-12
Also, does this have any performance impact?