@tothtamas28 2023-04-13
> Also, does this have any performance impact?

It makes KAST <-> JSON much faster, and KAST <-> KORE somewhat slower. Here are some measurements.

|              |   old |   new | speedup |
|--------------|-------|-------|---------|
| JSON -> KAST | 0.639 | **0.017** |     38x |
| KAST -> JSON | 0.594 | **0.002** |    297x |
| KAST -> KORE | **0.016** | 0.022 |    -30% |